refactor(gateway): GAR-476 — Q10.b extract build_channels to bootstrap/channels.rs
Slice 10.b of the bootstrap.rs modularisation (parent GAR-440). Extracts
`build_channels` (44 LOC) from `bootstrap/mod.rs` into a new sibling module
`bootstrap/channels.rs`. Re-exported at the same public path via
`pub use channels::build_channels` so all call sites remain unchanged.

bootstrap/mod.rs: 2358 → 2316 lines (−42).

No behaviour change. All imports verified:
  - `garraia_config::AppConfig` (parameter type)
  - `garraia_channels::ChannelRegistry` (return type + ChannelRegistry::new)
  - `dotenvy::dotenv` (full-path call, workspace dep of garraia-gateway)
  - `tracing::{info, warn, debug}` (logging macros)
